+++ CHANGELOG.md
@@ -10,8 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - enable mypy linting for tests folder
+- include persons from allowed primary sources in the publishing pipeline
 
 ### Changes
+
+- summarize logging for synopse filtering functions
 
 ### Deprecated